+++ cg_tpe2/src/main/java/ar/edu/itba/cg/tpe2/core/scene/OctreeNode.java
@@ -3,19 +3,19 @@ package ar.edu.itba.cg.tpe2.core.scene;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.vecmath.Point3d;
+import javax.vecmath.Point3f;
 
 import ar.edu.itba.cg.tpe2.core.geometry.Primitive;
 
 public class OctreeNode {
-	double xMin;
-	double xMax;
-	double yMin;
-	double yMax;
-	double zMin;
-	double zMax;
+	float xMin;
+	float xMax;
+	float yMin;
+	float yMax;
+	float zMin;
+	float zMax;
 	
-	double tMin;
+	float tMin;
 	
 	// Node's children
 	List<OctreeNode> childs = new ArrayList<OctreeNode>(8);
@@ -25,7 +25,7 @@ public class OctreeNode {
 	 * Default Constructor
 	 * 
 	 */
-	public OctreeNode(double xMin, double xMax, double yMin, double yMax, double zMin, double zMax) {
+	public OctreeNode(float xMin, float xMax, float yMin, float yMax, float zMin, float zMax) {
 		this.xMin=xMin;
 		this.yMin=yMin;
 		this.zMin=zMin;
@@ -38,7 +38,7 @@ public class OctreeNode {
 		return childs.isEmpty();
 	}
 	
-	public boolean contains(Point3d p) {
+	public boolean contains(Point3f p) {
 		if (p.x >= this.xMin && p.x < this.xMax && p.y >= this.yMin && p.y < this.yMax && p.z >= this.zMin && p.z < this.zMax) {
 			return true;
 		}
